refactor: move settings to _settings and update imports and parameter names
