+++ packages/kiarina-lib-google-auth/src/kiarina/lib/google/auth/__init__.py
@@ -6,12 +6,12 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from ._helpers.get_credentials import get_credentials
     from ._helpers.get_self_signed_jwt import get_self_signed_jwt
+    from ._settings import GoogleAuthSettings, settings_manager
     from ._types.credentials import Credentials
     from ._types.credentials_cache import CredentialsCache
     from ._utils.get_default_credentials import get_default_credentials
     from ._utils.get_service_account_credentials import get_service_account_credentials
     from ._utils.get_user_account_credentials import get_user_account_credentials
-    from .settings import GoogleAuthSettings, settings_manager
 
 __version__ = version("kiarina-lib-google-auth")
 
@@ -19,6 +19,9 @@ __all__ = [
     # ._helpers
     "get_credentials",
     "get_self_signed_jwt",
+    # ._settings
+    "GoogleAuthSettings",
+    "settings_manager",
     # ._types
     "Credentials",
     "CredentialsCache",
@@ -26,9 +29,6 @@ __all__ = [
     "get_default_credentials",
     "get_service_account_credentials",
     "get_user_account_credentials",
-    # .settings
-    "GoogleAuthSettings",
-    "settings_manager",
 ]
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
@@ -42,6 +42,9 @@ def __getattr__(name: str) -> object:
         # ._helpers
         "get_credentials": "._helpers.get_credentials",
         "get_self_signed_jwt": "._helpers.get_self_signed_jwt",
+        # ._settings
+        "GoogleAuthSettings": "._settings",
+        "settings_manager": "._settings",
         # ._types
         "Credentials": "._types.credentials",
         "CredentialsCache": "._types.credentials_cache",
@@ -49,9 +52,6 @@ def __getattr__(name: str) -> object:
         "get_default_credentials": "._utils.get_default_credentials",
         "get_service_account_credentials": "._utils.get_service_account_credentials",
         "get_user_account_credentials": "._utils.get_user_account_credentials",
-        # .settings
-        "GoogleAuthSettings": ".settings",
-        "settings_manager": ".settings",
     }
 
     globals()[name] = getattr(import_module(module_map[name], __name__), name)
